# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bakery ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with branch (optional), username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently logged in user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in

  # ---------- BRANCHES ----------
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create a manual attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in an employee for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out an employee for today (auto-computes hours, overtime, lateness)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "404":
          description: Not found
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------- LEAVES ----------
...
  /bonuses:
    get:
      operationId: listBonuses
      tags: [bonuses]
      summary: List bonuses
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bonus"
    post:
      operationId: createBonus
      tags: [bonuses]
      summary: Add a bonus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BonusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bonus"
  /bonuses/{id}:
    delete:
      operationId: deleteBonus
      tags: [bonuses]
      summary: Delete a bonus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------- DEDUCTIONS ----------
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [general_manager, branch_manager]
    User:
      type: object
      required: [id, username, fullName, role, branchId]
      properties:
        id: { type: string }
        username: { type: string }
        fullName: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        active: { type: boolean }
        createdAt: { type: string }
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        branchId:
          type: ["string", "null"]
        username: { type: string }
        password: { type: string }
...
    AttendanceStatus:
      type: string
      enum: [present, absent, leave, late]
    Attendance:
      type: object
      required: [id, employeeId, date, status]
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeName: { type: ["string", "null"] }
        branchId: { type: ["string", "null"] }
        date: { type: string }
        checkIn: { type: ["string", "null"] }
        checkOut: { type: ["string", "null"] }
        workMinutes: { type: ["integer", "null"] }
        overtimeMinutes: { type: ["integer", "null"] }
        lateMinutes: { type: ["integer", "null"] }
        earlyLeaveMinutes: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/AttendanceStatus"
    AttendanceInput:
      type: object
      required: [employeeId, date, status]
      properties:
        employeeId: { type: string }
        date: { type: string }
        checkIn: { type: string }
        checkOut: { type: string }
        notes: { type: string }
        status:
          $ref: "#/components/schemas/AttendanceStatus"
    AttendanceUpdate:
      type: object
      properties:
        checkIn: { type: string }
        checkOut: { type: string }
        notes: { type: string }
        status:
          $ref: "#/components/schemas/AttendanceStatus"
    CheckInInput:
      type: object
      required: [employeeId]
      properties:
        employeeId: { type: string }
    CheckOutInput:
      type: object
      required: [employeeId]
      properties:
        employeeId: { type: string }
...
    Bonus:
      type: object
      required: [id, employeeId, amount, reason, date]
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeName: { type: ["string", "null"] }
        amount: { type: number }
        reason: { type: string }
        date: { type: string }
    BonusInput:
      type: object
      required: [employeeId, amount, reason, date]
      properties:
        employeeId: { type: string }
        amount: { type: number, minimum: 0 }
        reason: { type: string, minLength: 1 }
        date: { type: string }
...